requirements.txt
langchain
langchain-google-genai
langchain-text-splitters
chromadb
pypdf
python-docx
odfpy
python-dotenv
pytest